+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/dto/GatheringPreview.java
@@ -1,9 +1,9 @@
 package com.triplem.momoim.core.domain.gathering.dto;
 
-import com.triplem.momoim.core.domain.gathering.Gathering;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
+import com.triplem.momoim.core.domain.gathering.model.Gathering;
 import com.triplem.momoim.core.domain.member.GatheringMemberDetail;
 import java.time.LocalDateTime;
 import java.util.List;
