+++ momoim-core/src/test/java/com/triplem/momoim/core/domain/gathering/GatheringBuilder.java
@@ -3,6 +3,7 @@ package com.triplem.momoim.core.domain.gathering;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
+import com.triplem.momoim.core.domain.gathering.model.Gathering;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Builder;
